requirements.txt
pyspark==3.5.1
requests==2.32.3
python-dotenv==1.0.1
boto3==1.34.162
